Firstweekbranch

+++ table.js
@@ -55,6 +55,7 @@ $(document).ready(function() {
       $(deleteBtn).click(function(e) {
         e.preventDefault();
         var url = $(this).attr("href");
+       if( confirm("Are you sure?")){
         $.ajax({
           type: "DELETE",
           url,
@@ -63,6 +64,7 @@ $(document).ready(function() {
             alert("Delete Successful");
           }
         });
+      }
       });
 
       $(editBtn).click(function(e) {
@@ -85,7 +87,9 @@ $(document).ready(function() {
             $("input[type='text']#timeIn").val(`${info.timeIn}`);
             $("input[type='text']#dateIn").val(`${info.dateIn}`);
             $("input[type='text']#hospNo").val(`${info.hospNo}`);
-            localStorage.userId = `${info.id}`;
+            $("#save").attr("value", info.id);
+           
+
           }
         });
       });
@@ -93,7 +97,7 @@ $(document).ready(function() {
 
     $("#ajax-form").submit(function(e) {
       e.preventDefault();
-      //var url = $("#ajax-form").attr("action");
+      var id = $("#save").attr("value");     
 
       var title = $("input[type='text']#title").val();
       var maritalStatus = $("input[type='text']#maritalStatus").val();
@@ -127,9 +131,12 @@ $(document).ready(function() {
 
       $.ajax({
         type: "PUT",
-        url: `http://localhost:3000/patients/${info.id}`,
+        url: `http://localhost:3000/PatientInfo/${
+          id
+        }`,
         data: modalObject,
         success: function() {
+          console.log(modalObject);
           alert("Patient Updated!");
           window.location.assign("table.html");
         }
